Remove Environment.exit() method
This method was needed for Python2 code where it was otherwise illegal to
`return` from a generator. Since Python3 supports `return` in generators,
the `exit()` method is unneeded. And since calling `env.exit()` is less
ergonomic than simply using `return`, there is no reason to keep it.

As an illustration, this generator written in Python2 style:

    def proc(env):
        yield env.timeout(1)
        env.exit(42)

can be expressed in Python3 as:

    def proc(env):
        yield env.timeout(1)
        return 42