Add inline type annotations
The primary goal for these annotations is to be helpful documentation for
users of the SimPy API. To this end, type annotations are added to each
SimPy method, function, and global. Also, by testing the value of
typing.TYPE_CHECKING at import time, the final, effective types for methods
bound with BoundClass magic can be expressed directly.

A secondary goal is to allow static checking for type soundness issues
using mypy. Mypy is added to requirements-dev.txt along with basic
configuration in setup.cfg. A few `# type: ignore` directives were needed
to deal with some of SimPy's internal optimizations, but with those in
place, SimPy type checks cleanly with mypy.

Special care was taken to avoid any performance regressions. Notably, a few
classes that might be expressed as generic using typing.Generic (e.g.
Event) were made non-generic to avoid the small-but-measurable performance
impact of instantiating a class that derives from Generic. Also, in a few
places we opted for using `# type: ignore` instead of adding an `assert
isinstance(...)`.