Event._value class attribute
This improves the runtime performance of Event and Process by decreasing
the initializer cost.

Adding the Event._value class attribute allows Event.__init__() to elide
initialization of the self._value instance attribute. For Event and
Process events, the event value is only assigned later in the event's
lifecycle. When self._value is eventually assigned, it creates an instance
attribute that shadows the static Event._value class attribute. This works
well for events because accesses to the Event._value class attribute are
rare/exceptional.