Add deploy stage to CI to upload to PyPI
The new deploy stage is limited to only running when the reference the
pipeline is running against is a tag.

A new twine-check job is added to the test stage. It checks the built
distribution files using `twine check`.

The twine upload command depends on the TWINE_USERNAME and TWINE_PASSWORD
environment variables being set by the CI runner. The username and password
should be a PyPI API token.

https://pypi.org/help/#apitoken