Build system updates for PEP-517/518
The distribution metadata is migrated from setup.py to setup.cfg. setup.py
is now just a stub. Note that `python setup.py ...` commands still work.

A pyproject.toml file is added. It contains the PEP-517/518 specified
build-system section which describes SimPy's build-time dependencies along
with which build backend SimPy uses (which is still setuptools).

And tox is now configured with isolated_build = True, which causes tox to
make the source dist in a sandbox instead of relying on the host
environment. This helps ensure that the distribution build will work with
just the tooling specified in pyproject.toml.